+++ pkg/app/src/pages/organisations/[...param].tsx
@@ -20,6 +20,7 @@ import {
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import { useTheme } from '@mui/material/styles'
+import { useAddMemberTransaction } from 'hooks/tx/useAddMemberTransaction'
 import { useConfig } from 'hooks/useConfig'
 import { useCurrentAccountAddress } from 'hooks/useCurrentAccountAddress'
 import { useTmpOrganisationState } from 'hooks/useTmpOrganisationState'
@@ -47,12 +48,18 @@ export function OrganisationById() {
 	const { loading, data } = useOrganizationByIdSubscription({
 		variables: { orgId: organizationIdState },
 	})
-
+	const [showTxModalType, setShowTxModalType] = useState<boolean>(false)
+	const addMemberTx = useAddMemberTransaction(organizationIdState)
 	const address = useCurrentAccountAddress()
 	const theme = useTheme()
 	const config = useConfig()
 	const tmpOrg = useTmpOrganisationState()
-
+	const handleOpenTxModal = useCallback(() => {
+		setShowTxModalType(true)
+	}, [setShowTxModalType])
+	const handleCloseTxModal = useCallback(() => {
+		setShowTxModalType(false)
+	}, [setShowTxModalType])
 	const isMd = useMediaQuery(theme.breakpoints.up('md'), {
 		defaultMatches: true,
 	})
@@ -257,7 +264,13 @@ export function OrganisationById() {
 										</Stack>
 										<Stack>
 											{!isMemberState && (
-												<Button variant="contained">{t('button:ui:join_organization')}</Button>
+												<Button
+													variant="contained"
+													disabled={!addMemberTx}
+													onClick={handleOpenTxModal}
+												>
+													{t('button:ui:join_organization')}
+												</Button>
 											)}
 										</Stack>
 									</Stack>
@@ -304,6 +317,10 @@ export function OrganisationById() {
 										organizationId={organizationIdState}
 										isMember={isMemberState}
 										isAdmin={address === organizationState?.controller}
+										handleCloseTxModal={handleCloseTxModal}
+										handleOpenTxModal={handleOpenTxModal}
+										showTxModalType={showTxModalType}
+										addMemberTx={addMemberTx}
 									/>
 								) : (
 									<TmpOverview />
